+++ docs/app/components/content/examples/composables/TextDemo.vue
@@ -3,7 +3,6 @@ import { stagger } from '#nanime/utils'
 import ExampleWrapper from '../../../shared/ExampleWrapper.vue'
 
 const text = useTemplateRef('phrase')
-
 const { chars } = useSplitText(text, {
   chars: true,
 })