+++ src/runtime/app/utils/normalizers/instance-management.ts
@@ -1,0 +1,11 @@
+import { getCurrentInstance } from 'vue'
+
+export enum AnimationComponentFlags {
+  Watchable, Instant,
+}
+
+export function getAnimationComponentFlag() {
+  return getCurrentInstance()
+    ? AnimationComponentFlags.Watchable
+    : AnimationComponentFlags.Instant
+};